完成8种博文操作

+++ src/main/java/com/duan/blogos/entity/blog/BlogAdmire.java
@@ -14,10 +14,20 @@ import java.sql.Timestamp;
 public class BlogAdmire implements Serializable {
 
     private static final long serialVersionUID = 4120356662820230266L;
+
+    //id
     private Integer id;
+
+    //博文id
     private Integer blogId;
+
+    //赞赏者id
     private Integer paierId;
-    private Double money;
+
+    //金额
+    private Float money;
+
+    //赞赏时间
     private Timestamp tranDate;
 
 }
